+++ src/ResultScan.php
@@ -26,14 +26,14 @@ class ResultScan extends Collection
      * Internal method to parse items.
      *
      * Ensures that each item is a `ScanEntity` and has all the properties it needs
-     * @param array|\Traversable $items Array of items
-     * @return array Parsed items
+     * @param iterable $items Items
+     * @return array<\LinkScanner\ScanEntity> Parsed items
      */
     protected function parseItems($items)
     {
         $items = $items instanceof Traversable ? $items->toArray() : $items;
 
-        return array_map(function ($item) {
+        return array_map(function ($item): ScanEntity {
             return $item instanceof ScanEntity ? $item : new ScanEntity($item);
         }, $items);
     }
@@ -46,6 +46,16 @@ class ResultScan extends Collection
     public function __construct($items = [])
     {
         parent::__construct($this->parseItems($items));
+    }
+
+    /**
+     * Unserializes the passed string and rebuilds the Collection instance
+     * @param string $collection The serialized collection
+     * @return void
+     */
+    public function unserialize($collection)
+    {
+        parent::__construct(unserialize($collection));
     }
 
     /**
@@ -67,7 +77,7 @@ class ResultScan extends Collection
      *
      * Returns a new `ResultScan` instance as the result of concatenating the
      *  passed list of elements with the list of elements in this collection
-     * @param array|\Traversable $items Items
+     * @param mixed $items The items to prepend
      * @return \Cake\Collection\CollectionInterface
      * @uses parseItems()
      */